Updated PreProcessor to:
a. set the PreProcessor P_INDEX values when a susceptiblity file is loaded as a spreadsheet, instead of a separate explicit call. (was causing widespread failure of tests before)

b. In PreProcessor, updated code to handle amino acid mixture substitution better. Previously, was substitution in the entire mixture (which was sometimes 2,3 or 4 in length.) Now, ONE of the amino acids in the mixture is randomly selected for substitution.